Implement dynamic indefinite articles and plurals
Analagous to {a} and {s} in Perchance.

These are done as a post-processing step rather than as part of a block,
since {s} in particular needs to look back and find the previous
instance of a number. This is less efficient, comprising two regex
passes over each finalized subpattern, but it does allow for a
convenient syntax.

Also, (s) can currently be used as a substitute for .pluralForm if there
is no number preceding it in the string. Nonetheless, an unconditional
.pluralForm equivalent will be implemented soon.