Generate a parse tree instead of re-parsing
This is a major overhaul of the engine. Now instead of parsing each
rule and query as a string from start to finish, each is tokenized and
saved as part of a parse tree. This means that the time complexity of
each pass over a string (after the initial tokenizing pass) is now
independent of the length of that string. Of course, it still depends on
the number of symbols and the depth of the tree, but each lookup
involved in generation should be constant time(?).

Other changes:

- Assigned values are now interpreted as special (block content) by
  default; the samples have been updated accordingly
- Rudimentary support for literals and patterns in blocks (cannot handle
  modifiers or surrounding whitespace yet)
- Modifiers are now just strings to keep things simple and consistent
- Reverted the previous (foolish) commit renaming `stripped` (`line` was
  still used in the function)